dictionary: Port to GSettings
Drop the GConf dependency for the main application, and use the new
GSettings API from GIO.

This requires some bumping around of dependencies, like GLib and GIO, to
something relatively recent.

The Dictionary window state is not stored in GSettings, but in a cache
file in XDG_CACHE_DIR/XDG_USER_CACHE, since it's something that we don't
need changes notification on, nor we particularly rely upon.